moved spin_rotate and spin_align to physics/sparse
This allows all spin-matrices to use spin-rotate and spin-align.

Fixes #920.
